LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/Create-non-absolute-url.any-expected.txt

CONSOLE MESSAGE: Wrong url scheme for WebSocket http://web-platform.test:8800/websockets/web-platform.test

PASS Create WebSocket - Pass a non absolute URL - SYNTAX_ERR is thrown
